Shaderfig: Add image fallbacks
After testing from my phone's browser and causing it to literally crash
the OS upon loading the page, I decided image fallbacks are a
requirement. Static images are now displayed until the user interacts
with the figure, at which point the shader is initialized and the image
is replaced with the live shader.

Images are generated with Selenium and headless Chrome, though headless
Firefox should also work. The browser must be headless in order to
specify screenshot dimensions precisely.